Implemented Question 9
Added functionality to check if a given number is an "Armstrong" number.
It counts the digits, then for each digit, it takes it to the power of
the total amount of digits, then adds them all together. If the result
is the same as the original number, it returns true.